feat: implement dynamic difficulty, extranonce, and improved block validation
Major improvements to stratum server functionality:

Dynamic Difficulty System:
- Add max_change_pct (0.5 default) to limit pool difficulty jumps per update
- Track template epoch from NNG pub/sub events
- Pass vardiff timing params through NetworkDifficultyTracker
- Adjust VarDiff clamp range to industry standard (0.67-1.5x)
- Update default min_difficulty from 4.0 to 0.5

Extranonce Support:
- Send mining.set_extranonce to miners on session start
- Use standard extranonce2_size=4 (2^32 nonces per extranonce1)
- Properly integrate extranonce1/extranonce2 into coinbase construction

Block Validation & Hash Computation:
- Add build_candidate_block_with_stratum_hash() returning hash + merkle root
- Pass height, epoch_hash, extended_metadata_hash, size to build_stratum_header
- Compute stratum hash matching GPU miner computation and lotusd validation
- Add detailed header/template debug logging for share acceptance
- Update block height type from i64 to i32 throughout

MiningJob Extensions:
- Include block_height, epoch_hash_hex, extended_metadata_hash_hex in notify params
- Extract epoch/extended metadata hashes from template block header

Documentation:
- Expand README with dynamic difficulty explanation, VarDiff algorithm details
- Document PPLNS payout system, transaction signing, deployment guide
- Add troubleshooting, monitoring, security considerations
- Update config.example.toml with comprehensive vardiff documentation

Configuration:
- Update share_target_ratio default to 50M for testnet compatibility
- Set sensible vardiff defaults: target=15s, retarget=90s

+++ src/accounting/sqlite.rs
@@ -554,7 +554,7 @@ impl AccountingDb {
         &self,
         block_hash: &str,
         template_id: u64,
-        block_height: i64,
+        block_height: i32,
         worker_id: i64,
         worker_name: &str,
         payout_address: &str,
